+++ varfish-annotator-cli/src/main/java/com/github/bihealth/varfish_annotator/annotate_svs/AnnotateSvsArgs.java
@@ -35,8 +35,8 @@ public final class AnnotateSvsArgs {
   @Parameter(names = "--release", description = "The genome release used", required = true)
   private String release;
 
-  @Parameter(names = "--case-id", description = "The value to use for case ID", required = true)
-  private String caseId;
+  @Parameter(names = "--set-id", description = "The value to use for set ID", required = true)
+  private String setId;
 
   @Parameter(
       names = "--input-vcf",
@@ -87,8 +87,8 @@ public final class AnnotateSvsArgs {
     return release;
   }
 
-  public String getCaseId() {
-    return caseId;
+  public String getSetId() {
+    return setId;
   }
 
   public String getInputVcf() {
@@ -128,8 +128,8 @@ public final class AnnotateSvsArgs {
         + ", release='"
         + release
         + '\''
-        + ", caseId='"
-        + caseId
+        + ", setId='"
+        + setId
         + '\''
         + ", inputVcf='"
         + inputVcf